==> 1_code/r_notebooks/1_ibutton_data_prepare.Rmd <==
```{r setup, include=FALSE, cache=FALSE}
#Set root directory to R project root
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

setwd(rprojroot::find_rstudio_root_file())
``` 

```{r eval=FALSE}
library(tidyr)
library(plyr)
library(dplyr)
# library(eExtra)
library(readxl)
library(lubridate)
library(imputeTS)
library(readr)
```

### Import and clean {.unnumbered}

#### RIVR {.unnumbered}

**Load data**

```{r}
RIVR<-read.csv(file="0_data/external/iButton/RIVR/iButtons_RIVR_combined_April7_2022_no_extremes.csv")
```

**Examine data frame**

Count the number of unique iButtons

```{r echo=FALSE, message = FALSE, results="asis"}
# count unique deployments
nrow(dplyr::distinct(as.data.frame((RIVR$Site_StationKey))))
```

View data

```{r echo=FALSE, message = FALSE, results="asis"}
#RIVR<-read.csv(file="0_data/external/iButton/RIVR/iButtons_RIVR_combined_April7_2022_no_extremes.csv")
knitr::kable(head(RIVR), "pipe") 
```

##### Edit data columns {.unnumbered}

###### Format date-time string {.unnumbered}

Convert date-time string into a `POSIXct` class. This class associates the date time string with an associated time zone. Codes for time zones can be found [here](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones). Else the `as.POSIXct` function defaults to your computer's timezone.

```{r eval=FALSE}
RIVR$Date_Time<-as.POSIXct(RIVR$Date_Time, tz="America/Edmonton")
```

##### Remove pre-deployement data {.unnumbered}

```{r eval=FALSE}
RIVR_2<- RIVR%>%
  filter(Date_Time>Date_Time_dpl)%>%
  filter(Date_Time<Date_Time_rtv)
```


##### Get daily temperature summaries {.unnumbered}

Create new calculated columns with the mean, max, and min daily temperatures and diurnal range. 

```{r eval=FALSE}
RIVR_dailys<-RIVR_2 %>% 
  mutate(Date=date(Date_Time))%>%
  dplyr::group_by(Site_StationKey,Day,Month,Year,Date, iBt_type)%>%  dplyr::mutate(Temperature=Value) %>%
  dplyr::summarize(Tmax_Day=max(Temperature),Tmin_Day=min(Temperature),Tavg_Day=mean(Temperature))%>%
  dplyr::group_by(Site_StationKey,iBt_type,Month,Year) %>% dplyr::filter(!iBt_type=="EXTRA-TOP")%>%
  arrange(Site_StationKey, Date)

save(RIVR_dailys, file="2_pipeline/store/RIVR_dailys.rData")
```



```{r echo=FALSE, message = FALSE, results="asis"}
load("2_pipeline/store/RIVR_dailys.rData")

knitr::kable(head(RIVR_dailys), "pipe") 
```


#### HILL {.unnumbered}

**Load data**

```{r}
hills<-read.csv(file="0_data/external/iButton/Hills/Hills_iButton_Data_Combined_Corrected_for_Deployment_no_extremes_Apr_27.csv")
```

**Examine data frame**

Count the number of unique iButtons

```{r echo=FALSE, message = FALSE, results="asis"}
# count unique deployments
nrow(dplyr::distinct(as.data.frame((hills$Site_StationKey))))
```

View data

```{r echo=FALSE, message = FALSE, results="asis"}
#RIVR<-read.csv(file="0_data/external/iButton/RIVR/iButtons_RIVR_combined_April7_2022_no_extremes.csv")
knitr::kable(head(hills), "pipe") 
```

##### Edit data columns {.unnumbered}

###### Format date-time string {.unnumbered}

Convert date-time string into a `POSIXct` class. This class associates the date time string with an associated time zone. Codes for time zones can be found [here](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones). Else the `as.POSIXct` function defaults to your computer's timezone.

```{r eval=FALSE}
hills$Date.Time<-as.POSIXct(hills$Date.Time, tz="America/Edmonton")
```


##### Get daily temperature summaries {.unnumbered}

Create new calculated columns with the mean, max, and min daily temperatures.

```{r eval=FALSE}
hills_dailys<-hills %>%
   mutate(Value=Temperature,Month=month(Date.Time),Day=day(Date.Time),Year=year(Date.Time),
                Date_Time=Date.Time) %>%
  mutate(Date=date(Date_Time))%>%
  dplyr::group_by(Site_StationKey,Date) %>% #group by days, month and year
  dplyr::summarize(Tmax_Day=max(Temperature),Tmin_Day=min(Temperature),Tavg_Day=mean(Temperature))%>%
  mutate(Year=year(Date))%>%
  mutate(Month=month(Date))%>%
  mutate(Day=day(Date))%>%
  dplyr::group_by(Site_StationKey,Date) 

save(hills_dailys, file="2_pipeline/store/hills_dailys.rData")
```

```{r echo=FALSE, message = FALSE, results="asis"}
load("2_pipeline/store/hills_dailys.rData")

knitr::kable(head(hills_dailys), "pipe") 
```


#### Wood {-} 

Data from Wood, Wendy H; Marshall, Shawn J; Fargey, Shannon E; Whitehead, Terri L (2017): Daily temperature data from the Foothills Climate Array Mesonet, Canadian Rocky Mountains, 2005-2010. PANGAEA, https://doi.org/10.1594/PANGAEA.880611. 

```{r eval=FALSE}
wood_dailys<-read.delim(file="0_data/external/iButton/WoodEtAl/Wood_etal_2017.tab",  header=T, sep="\t", skip = 23)%>%
  rename(c(Site_StationKey=Station..FCAID., Date_Time=Date.Time, Tmin_Day=TnTnTn.day.min...C., Tmax_Day=TxTxTx.day.max...C., Tavg_Day=TTT.daily.m...C.))%>%
  select(Site_StationKey, Date_Time, Tmin_Day, Tmax_Day, Tavg_Day)%>%
  mutate(Date=date(Date_Time))%>%
  mutate(Year=year(Date))%>%
  mutate(Month=month(Date))%>%
  mutate(Day=day(Date))%>%
  select(-Date_Time)

save(wood_dailys, file="2_pipeline/store/wood_dailys.rData")
```

#### ARVTDendro {-}

```{r}
load("0_data/external/iButton/ARVT/ARVTDendro/arvtdendro2021_cleaned_iButton_data.RData")

ARVTDendro_2021_dailys<-mk_cleaned%>%
  mutate(Value=Temperature, Site_StationKey=mission_id)%>%
  mutate(Date=date(Date_Time))%>%
  filter(Date_Time>Date_Time_dpl)%>%
  filter(Date<=retr_date)%>%
  dplyr::group_by(Site_StationKey,Date) %>% #group by days, month and year
  dplyr::summarize(Tmax_Day=max(Temperature),Tmin_Day=min(Temperature),Tavg_Day=mean(Temperature))%>%
  mutate(Year=year(Date))%>%
  mutate(Month=month(Date))%>%
  mutate(Day=day(Date))%>%
  dplyr::group_by(Site_StationKey,Date) 

save(ARVTDendro_2021_dailys, file="2_pipeline/store/ARVTDendro_2021_dailys.rData")
```

#### ARVTFire {-}

```{r}
load("0_data/external/iButton/ARVT/ARVTFire/alex_2020_2021_cleaned_temp_data.RData")
ARVTFire_2020_2021_dailys<-alex_2020_cleaned%>%
  mutate(Value=Temperature, Site_StationKey=mission_id)%>%
  mutate(Date=date(Date_Time))%>%
  filter(Date_Time>date_time_deployed)%>%
  filter(Date<=retr_date_time)%>%
  dplyr::group_by(Site_StationKey,Date) %>% #group by days, month and year
  dplyr::summarize(Tmax_Day=max(Temperature),Tmin_Day=min(Temperature),Tavg_Day=mean(Temperature))%>%
  mutate(Year=year(Date))%>%
  mutate(Month=month(Date))%>%
  mutate(Day=day(Date))%>%
  dplyr::group_by(Site_StationKey,Date) 

load("0_data/external/iButton/ARVT/ARVTFire/alex_2021_2022_cleaned_temp_data.RData")
ARVTFire_2021_2022_dailys<-alex_processed%>%
  mutate(Value=Temperature, Site_StationKey=mission_id)%>%
  mutate(Date=date(Date_Time))%>%
  filter(Date_Time>date_time_deployed)%>%
  filter(Date<=retr_date_time)%>%
  dplyr::group_by(Site_StationKey,Date) %>% #group by days, month and year
  dplyr::summarize(Tmax_Day=max(Temperature),Tmin_Day=min(Temperature),Tavg_Day=mean(Temperature))%>%
  mutate(Year=year(Date))%>%
  mutate(Month=month(Date))%>%
  mutate(Day=day(Date))%>%
  dplyr::group_by(Site_StationKey,Date) 

ARVTFire_2020_2022_dailys<-rbind(ARVTFire_2021_2022_dailys, ARVTFire_2020_2021_dailys)

save(ARVTFire_2020_2022_dailys, file="2_pipeline/store/ARVTFire_2020_2022_dailys.rData")
```


#### Red Plateau {-}

The following data is from Derek van der Kamp

```{r}
RedPlat<-read.csv(file="0_data/external/iButton/Derek_VanDerKamp/redplateau.met.csv")

RedPlat_dailys<-RedPlat%>%
  mutate(Temperature=value)%>%
  mutate(Site_StationKey=str_c(site, subsite, sep="_"))%>%
  filter(name=="airtemp")%>%
  mutate(Date=date(time))%>%
  dplyr::group_by(Site_StationKey,Date) %>% #group by days, month and year
  dplyr::summarize(Tmax_Day=max(Temperature),Tmin_Day=min(Temperature),Tavg_Day=mean(Temperature))%>%
  mutate(Year=year(Date))%>%
  mutate(Month=month(Date))%>%
  mutate(Day=day(Date))%>%
  dplyr::group_by(Site_StationKey,Date)

save(RedPlat_dailys, file="2_pipeline/store/RedPlat_dailys.rData")
```


#### BC Weather station {-}

BC weather station data from https://bcgov-env.shinyapps.io/nbcclim/

```{r}
# from https://bcgov-env.shinyapps.io/nbcclim/:
BCWS<-read.csv(file="0_data/external/iButton/BC_weatherStations/All_stations.csv")



BCWS_dailys<-BCWS%>%
  mutate(Site_StationKey=Site)%>%
  mutate(Tmax_Day=Temp_max)%>%
  mutate(Tmin_Day=Temp_min)%>%
  mutate(Tavg_Day=Temp_avg)%>%
  mutate(Date=as.Date(Date))%>%
  mutate(Year=year(Date))%>%
  mutate(Month=month(Date))%>%
  mutate(Day=day(Date))%>%  
  select(Site,Day, Month, Year, Date, Tmax_Day, Tmin_Day, Tavg_Day)%>%
  na.omit()

save(BCWS_dailys, file="2_pipeline/store/BCWS_dailys.rData")
```

#### Cardinal Divide {-}

```{r eval=FALSE}

# Bring in meta data and keep on;y ibuttons that are 1.5 vm off the ground
crdv_met_1<-read_excel("0_data/external/iButton/cardinal_divide/cardinaldividetemperaturestationsdata/Nielsen_02_SoilTemp2.0_data-submission_template_long_unique_sensor.xlsx", sheet = "Metadata")

crdv_met_2<-read_excel("/Volumes/Projects/climate_downscaling/0_data/external/iButton/cardinal_divide/fwcardinaldivideelevationtransecttemperaturedata/Nielsen-SoilTemp2.0_data-submission_template_long_unique_sensor.xlsx", sheet = "Metadata")

crdv_met_all<-crdv_met_1%>%
  bind_rows(crdv_met_2)%>%
  select(Raw_data_identifier, Sensor_height)%>%
  rename(Site_StationKey=Raw_data_identifier)

# Read raw data
crdv_1<-read_excel("0_data/external/iButton/cardinal_divide/cardinaldividetemperaturestationsdata/Nielsen_02_SoilTemp2.0_data-submission_template_long_unique_sensor.xlsx", sheet = "Raw time series")

crdv_2<-read_excel("/Volumes/Projects/climate_downscaling/0_data/external/iButton/cardinal_divide/fwcardinaldivideelevationtransecttemperaturedata/Nielsen-SoilTemp2.0_data-submission_template_long_unique_sensor.xlsx", sheet = "Raw time series")

crdv_dailys<-crdv_1%>%
  bind_rows(crdv_2)%>%
  rename(c(time_24h=`Time (24h)`, Site_StationKey=Raw_data_identifier))%>%
  left_join(crdv_met_all)%>%
  filter(Sensor_height==150)%>%
  mutate(time_24h=format(time_24h, format = "%H:%M"))%>%
  dplyr::group_by(Site_StationKey, Year, Month, Day) %>% #group by days, month and year
  dplyr::summarize(Tmax_Day=max(Temperature),Tmin_Day=min(Temperature),Tavg_Day=mean(Temperature)) %>%
  mutate(Date = ymd(paste(Year, Month, Day, sep = "-")))%>%
  select(Site_StationKey, Date, Tmax_Day, Tmin_Day, Tavg_Day, Year, Month, Day)


save(crdv_dailys, file="2_pipeline/store/crdv_dailys.rData")
```


### Combine datasets {.unnumbered}

#### **Bind dataframes**

```{r eval=FALSE}

load("2_pipeline/store/hills_dailys.rData")
load("2_pipeline/store/RIVR_dailys.rData")
load("2_pipeline/store/wood_dailys.rData")
load("2_pipeline/store/ARVTFire_2020_2022_dailys.rData")
load("2_pipeline/store/ARVTDendro_2021_dailys.rData")
load("2_pipeline/store/RedPlat_dailys.rData")
load("2_pipeline/store/BCWS_dailys.rData")
load("2_pipeline/store/crdv_dailys.rData")

ibuttons<-bind_rows("RIVR"=RIVR_dailys,"HILLS"=hills_dailys,  "ARVTFire"=ARVTFire_2020_2022_dailys, "ARVTDendro"=ARVTDendro_2021_dailys, "wood"=wood_dailys, "RedPlat"=RedPlat_dailys, "BCWS"=BCWS_dailys, "cardinal_divide"=crdv_dailys, .id="Project")

save(ibuttons, file=paste0("2_pipeline/tmp/ibuttons_", Sys.Date(), ".rData"))
```
