==> 1_code/r_notebooks/2_ibutton_data_xy.Rmd <==
```{r setup, include=FALSE, cache=FALSE}
#Set root directory to R project root
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```
 


```{r eval=FALSE}
library(sf)
library(tmap)
library(basemaps)
library(dplyr)
library(kableExtra)
library(readxl)
library(lubridate)
library(purrr)
```

### RIVR {.unnumbered}

**Load data**

```{r eval=FALSE}
RIVR<-read.csv(file="0_data/external/iButton/RIVR/iButtons_RIVR_combined_April7_2022_no_extremes.csv")
```


#### Create spatial data frame of iButton locations {.unnumbered}

Extract XY coordinates and save as shapefile.

```{r eval=FALSE}
RIVR_xy<-RIVR%>%
  group_by(Site_StationKey)%>%
  mutate(min_year=min(Year))%>%
  mutate(max_year=max(Year))%>%
  dplyr::select(c(Project, Site_StationKey, Lat, Long, min_year, max_year))%>%
  dplyr::distinct()

RIVR_xy<-st_as_sf(RIVR_xy, coords=c("Long","Lat"), crs=4326, remove=FALSE)

# save as spatial data frame
save(RIVR_xy, file="0_data/manual/iButton_data/spatial/RIVR_xy.rData")

# save as shapefile
st_write(RIVR_xy, "0_data/manual/iButton_data/spatial/RIVR_xy.shp")
```


### HILL {.unnumbered}

**Load data**

```{r eval=FALSE}
hills<-read.csv(file="0_data/external/iButton/Hills/Hills_iButton_Data_Combined_Corrected_for_Deployment_no_extremes_Apr_27.csv")

hills_loc<-read_xlsx("0_data/external/iButton/Hills/SiteLocations/Hl_coordinates.xlsx")

load("2_pipeline/store/hills_dailys.rData")
```

#### Create spatial data frame of iButton locations {.unnumbered}

Extract XY coordinates and save as shapefile.

```{r eval=FALSE}
# create a station key column and convert to spatial object
hills_loc_2<-hills_loc%>%
      mutate(Site_StationKey=paste(hills_loc$ProjectID, hills_loc$Cluster, hills_loc$SITE, hills_loc$STATION, sep="-"))%>%
      inner_join(hills_dailys)%>%
      dplyr::rename(Project=ProjectID, Date_deplo=DEPLOY_DATE)%>%
      group_by(Site_StationKey)%>%
      mutate(min_year=min(Year))%>%
      mutate(max_year=max(Year))%>%
      mutate(Project="HILLS")%>%
      # left_join(test, by=c("EASTING"="X", "NORTHING"="Y"))%>%
      dplyr::select(c(Project, Site_StationKey, min_year, max_year, EASTING, NORTHING, UTM_Zone))%>%
      distinct()

hills_xy_zone11<-hills_loc_2%>%
      filter(UTM_Zone==11)%>%
      st_as_sf(coords=c("EASTING","NORTHING"), crs=26911)%>% #NAD83 / UTM zone 11N
      st_transform(crs=st_crs(4326))%>%
      mutate(Long = unlist(map(geometry,1)),
           Lat = unlist(map(geometry,2)))%>%
      select(c(Project, Site_StationKey, Lat, Long, min_year, max_year, geometry))


hills_xy_zone12<-hills_loc_2%>%
      filter(UTM_Zone==12)%>%
      st_as_sf(coords=c("EASTING","NORTHING"), crs=26912)%>% #NAD83 / UTM zone 11N
      st_transform(crs=st_crs(4326))%>%
      mutate(Long = unlist(map(geometry,1)),
           Lat = unlist(map(geometry,2)))%>%
      select(c(Project, Site_StationKey, Lat, Long, min_year, max_year, geometry))


hills_xy<-rbind(hills_xy_zone11, hills_xy_zone12)



# save as spatial data frame
save(hills_xy, file="0_data/manual/iButton_data/spatial/hills_xy.rData")

# save as shapefile
st_write(hills_xy, "0_data/manual/iButton_data/spatial/hills_xy.shp", delete_dsn = TRUE)
```


### Wood {.unnumbered}

**Load data**

```{r eval=FALSE}
wood<-read.delim(file="0_data/external/iButton/WoodEtAl/Wood_etal_2017.tab",  header=T, sep="\t", skip = 23)

```

#### Create spatial data frame of iButton locations {.unnumbered}

Extract XY coordinates and save as shapefile.

```{r eval=FALSE}
wood_xy<-wood%>%
  mutate(Site_StationKey=Station..FCAID.)%>%
  mutate(Project="wood")%>%
  group_by(Site_StationKey)%>%
  mutate(min_year=min(year(Date.Time)))%>%
  mutate(max_year=max(year(Date.Time)))%>%
  dplyr::select(c(Project, Site_StationKey, Latitude, Longitude, min_year, max_year))%>%
  rename(c(Lat=Latitude, Long=Longitude))%>%
  dplyr::distinct()

wood_xy<-st_as_sf(wood_xy, coords=c("Long","Lat"), crs=4326, remove=FALSE)

# save as spatial data frame
save(wood_xy, file="0_data/manual/iButton_data/spatial/wood_xy.rData")

# save as shapefile
st_write(wood_xy, "0_data/manual/iButton_data/spatial/wood_xy.shp")
```

### Alex {.unnumbered}

**Load data**

```{r eval=FALSE}
load("2_pipeline/store/alex_dailys.rData")

# bring in xy data
alex_loc<-read_csv(file="0_data/external/iButton/old/alex/2019_and_2020_2021_plot_locations_and_elevations.csv")
```


#### Create spatial data frame of iButton locations {.unnumbered}

Extract XY coordinates and save as shapefile.

```{r eval=FALSE}
alex_xy<-alex_loc%>%
  mutate(Site_StationKey=paste(site, plot, sep="_"))%>%
  inner_join(alex_dailys)%>%
  group_by(Site_StationKey)%>%
  mutate(min_year=min(Year))%>%
  mutate(max_year=max(Year))%>%
  mutate(Project="alex")%>%
  dplyr::select(c(Project, Site_StationKey, lat, long, min_year, max_year))%>%
  rename(c(Lat=lat, Long=long))%>%
  dplyr::distinct()

alex_xy<-st_as_sf(alex_xy, coords=c("Long","Lat"), crs=4326, remove=FALSE)


# save as spatial data frame
save(alex_xy, file="0_data/manual/iButton_data/spatial/alex_xy.rData")

# save as shapefile
st_write(alex_xy, "0_data/manual/iButton_data/spatial/alex_xy.shp")
```




### ARVTDendro {-}

```{r}
load("0_data/external/iButton/ARVT/ARVTDendro/arvtdendro2021_lat_longs.RData")

ARVTDendro_2021_xy<-mk_lat_long%>%
  mutate(Site_StationKey=mission_id)%>%
  mutate(Project="ARVTDendro")%>%
  group_by(Site_StationKey)%>%
  mutate(min_year=min(year(Date_Time_dpl)))%>%
  mutate(max_year=max(year(retr_date)))%>%
  rename(c(Lat=lat, Long=long))%>%
  mutate(Lat=as.numeric(Lat))%>%
  mutate(Long=as.numeric(Long))%>%
  dplyr::select(c(Project, Site_StationKey, Lat, Long, min_year, max_year))%>%
  dplyr::distinct()%>%
  st_as_sf(coords=c("Long","Lat"), crs=4326, remove=FALSE)

# save as spatial data frame
save(ARVTDendro_2021_xy, file="0_data/manual/iButton_data/spatial/ARVTDendro_2021_xy.rData")

# save as shapefile
st_write(ARVTDendro_2021_xy, "0_data/manual/iButton_data/spatial/ARVTDendro_2021_xy.shp")
```

### ARVTFire {-}

```{r}
load("0_data/external/iButton/ARVT/ARVTFire/alex_2020_2021_cleaned_temp_data.RData")
ARVTFire2021_locations_comments <-read_csv("0_data/external/iButton/ARVT/ARVTFire/ARVTFire2021_locations_comments.csv")

ARVTFire_2020_2021_xy<-alex_2020_cleaned%>%
  mutate(Site_StationKey=mission_id)%>%
  mutate(Project="ARVTFire")%>%
  group_by(Site_StationKey)%>%
  mutate(min_year=min(year(Date_Time)))%>%
  mutate(max_year=max(year(Date_Time)))%>%
  dplyr::select(c(Project, Site_StationKey,min_year, max_year))%>%
  distinct()%>%
  left_join(ARVTFire2021_locations_comments,by = c("Site_StationKey"="mission_id"))%>%
  rename(c(Lat=Latitude, Long=Longitude))%>%
  dplyr::select(c(Project, Site_StationKey, Lat, Long, min_year, max_year))%>%
  dplyr::distinct()

load("0_data/external/iButton/ARVT/ARVTFire/alex_2021_2022_cleaned_temp_data.RData")
ARVTFire_2021_2022_xy<-alex_processed%>%
  mutate(Site_StationKey=mission_id)%>%
  mutate(Project="ARVTFire")%>%
  group_by(Site_StationKey)%>%
  mutate(Date=date(Date_Time))%>%
  mutate(min_year=min(year(Date)))%>%
  mutate(max_year=max(year(Date)))%>%
  rename(c(Lat=lat, Long=long))%>%
  dplyr::select(c(Project, Site_StationKey, Lat, Long, min_year, max_year))%>%
  dplyr::distinct()

ARVTFire_2020_2022_xy<-rbind(ARVTFire_2021_2022_xy, ARVTFire_2020_2021_xy)%>%
  na.omit()%>%
  st_as_sf(coords=c("Long","Lat"), crs=4326, remove=FALSE)

# save as spatial data frame
save(ARVTFire_2020_2022_xy, file="0_data/manual/iButton_data/spatial/ARVTFire_2020_2022_xy.rData")

# save as shapefile
st_write(ARVTFire_2020_2022_xy, "0_data/manual/iButton_data/spatial/ARVTFire_2020_2022_xy.shp")
```



### Red Plateau {-}

The following data is from Derek van der Kamp

```{r}
RedPlat<-read.csv(file="0_data/external/iButton/Derek_VanDerKamp/redplateau.met.csv")

RedPlat_xy<-RedPlat%>%
    mutate(Site_StationKey=str_c(site, subsite, sep="_"))%>%
    mutate(Project="RedPlat")%>%
    group_by(Site_StationKey)%>%
    mutate(Date=date(time))%>%
    mutate(min_year=min(year(Date)))%>%
    mutate(max_year=max(year(Date)))%>%
    rename(c(Lat=lat, Long=lon))%>%
    mutate(Lat=as.numeric(Lat))%>%
    mutate(Long=as.numeric(Long))%>%
    dplyr::select(c(Project, Site_StationKey, Lat, Long, min_year, max_year))%>%
    dplyr::distinct()%>%
    st_as_sf(coords=c("Long","Lat"), crs=4326, remove=FALSE)
  
save(RedPlat_xy, file="0_data/manual/iButton_data/spatial/RedPlat_xy.rData")

# save as shapefile
st_write(RedPlat_xy, "0_data/manual/iButton_data/spatial/RedPlat_xy.shp")
  
```


### BC Weather station {-}

BC weather station data from https://bcgov-env.shinyapps.io/nbcclim/

```{r}
BCWS<-read.csv(file="0_data/external/iButton/BC_weatherStations/All_stations.csv")

BCWS_xy<-BCWS%>%
  mutate(Site_StationKey=Site)%>%
  mutate(Project="BCWS")%>%
  group_by(Site_StationKey)%>%
  mutate(Date=as.Date(Date))%>%
  mutate(min_year=min(year(Date)))%>%
  mutate(max_year=max(year(Date)))%>%
  rename(c(Lat=Latitude, Long=Longitude))%>%
  mutate(Lat=as.numeric(Lat))%>%
  mutate(Long=as.numeric(Long))%>%
  dplyr::select(c(Project, Site_StationKey, Lat, Long, min_year, max_year))%>%
  dplyr::distinct()%>%
  st_as_sf(coords=c("Long","Lat"), crs=4326, remove=FALSE)

save(BCWS_xy, file="0_data/manual/iButton_data/spatial/BCWS_xy.rData")

# save as shapefile
st_write(BCWS_xy, "0_data/manual/iButton_data/spatial/BCWS_xy.shp")
```


### Cardinal Divide {-}

```{r eval=FALSE}
crdv_met_1<-read_excel("0_data/external/iButton/cardinal_divide/cardinaldividetemperaturestationsdata/Nielsen_02_SoilTemp2.0_data-submission_template_long_unique_sensor.xlsx", sheet = "Metadata")

crdv_met_2<-read_excel("/Volumes/Projects/climate_downscaling/0_data/external/iButton/cardinal_divide/fwcardinaldivideelevationtransecttemperaturedata/Nielsen-SoilTemp2.0_data-submission_template_long_unique_sensor.xlsx", sheet = "Metadata")

crdv_xy<-crdv_met_1%>%
  bind_rows(crdv_met_2)%>%
  slice(-1)%>%
  select(Raw_data_identifier, Latitude, Longitude, Start_date_year,End_date_year)%>%
  rename(c(min_year=Start_date_year, max_year=End_date_year, Site_StationKey=Raw_data_identifier))%>%
  rename(c(Lat=Latitude, Long=Longitude))%>%
  mutate(Lat=as.numeric(Lat))%>%
  mutate(Lon=as.numeric(Long))%>%
  mutate(Project="CRDV")%>%
  dplyr::select(c(Project, Site_StationKey, Lat, Long, min_year, max_year))%>%
  dplyr::distinct()%>%
  na.omit()%>%
  st_as_sf(coords=c("Long","Lat"), crs=4326, remove=FALSE)

save(crdv_xy, file="0_data/manual/iButton_data/spatial/crdv_xy.rData")

# save as shapefile
st_write(crdv_xy, "0_data/manual/iButton_data/spatial/crdv_xy.shp")
```

### BU projects

```{r}
BUMTN2022_xy <- st_read("~/Google_Drive/Shared drives/CFS_microclimate/Projects/BUMTN2022/cleaned/BUMTN2022_xy_20231016.shp")
BUGhost2022_xy <- load("~/Google_Drive/Shared drives/CFS_microclimate/Projects/BUGhost2022/cleaned/BUGhost2022_xy_20231016.rData")
BUCL2022_xy <- load("~/Google_Drive/Shared drives/CFS_microclimate/Projects/BUCL2022/cleaned/BUCL2022_xy_20231016.rData")
BUGen2022_xy <- load("~/Google_Drive/Shared drives/CFS_microclimate/Projects/BUGen2022/cleaned/BUGen2022_xy_20231016.rData")

BU_xy<-rbind(BUCL2022_xy, BUGhost2022_xy, BUMTN2022_xy, BUGen2022_xy)

%>%
  select(project_id, mission_id, latitude, longitude, year)%>%
    group_by(project_id, mission_id, latitude, longitude)%>%
    dplyr::summarize(min_year=min(year), max_year=max(year))%>%
    dplyr::rename(c(Project=project_id, Site_StationKey=mission_id, Lat=latitude, Long=longitude))%>%
    dplyr::distinct()%>%
    na.omit()%>%
    st_as_sf(coords=c("Long","Lat"), crs=4326, remove=FALSE)


st_write(BU_xy, "0_data/manual/iButton_data/spatial/BU_xy.shp")
  


  

```





###






### All projects {-}

```{r eval=FALSE}
load("0_data/manual/iButton_data/spatial/BCWS_xy.rData")
load("0_data/manual/iButton_data/spatial/RIVR_xy.rData")
load("0_data/manual/iButton_data/spatial/hills_xy.rData")
load("0_data/manual/iButton_data/spatial/wood_xy.rData")
load("0_data/manual/iButton_data/spatial/alex_xy.rData")
load("0_data/manual/iButton_data/spatial/ARVTDendro_2021_xy.rData")
load("0_data/manual/iButton_data/spatial/ARVTFire_2020_2022_xy.rData")
load("0_data/manual/iButton_data/spatial/RedPlat_xy.rData")
load("0_data/manual/iButton_data/spatial/crdv_xy.rData")

BUGhost2022_xy<-readRDS("~/Google Drive/Shared drives/CFS_microclimate/Projects/BUGhost2022/cleaned/BUGhost2022_xy_20231017.rds")
BUCL2022_xy<-readRDS("~/Google Drive/Shared drives/CFS_microclimate/Projects/BUCL2022/cleaned/BUCL2022_xy_20231017.rds")
BUGen2022_xy<-readRDS("~/Google Drive/Shared drives/CFS_microclimate/Projects/BUGen2022/cleaned/BUGen2022_xy_20231017.rds")
BUMTN2022_xy<-readRDS("~/Google Drive/Shared drives/CFS_microclimate/Projects/BUMTN2022/cleaned/BUMTN2022_xy_20231017.rds")


BU_xy<-rbind(BUCL2022_xy, BUGhost2022_xy, BUMTN2022_xy, BUGen2022_xy)

ss_xy<-rbind(RIVR_xy, wood_xy, hills_xy, alex_xy, BCWS_xy, RedPlat_xy, ARVTDendro_2021_xy, ARVTFire_2020_2022_xy, crdv_xy, BU_xy)

# save as spatial data frame
save(ss_xy, file="0_data/manual/iButton_data/spatial/ss_xy.rData")

# save as shapefile
st_write(ss_xy, "0_data/manual/iButton_data/spatial/ss_xy.shp", delete_dsn = TRUE)
```


### Map study area {-}

#### Identify study area {.unnumbered}

```{r eval=FALSE}
#create a bounding box around study area
bb<-st_bbox(ss_xy)

#Get aspect ratio of bounding box
bb<-st_as_sfc(bb)
bb<-st_as_sf(bb)
bb_buf<-st_buffer(bb, 10000)
bb_buf<-st_bbox(bb_buf)
bb_buf<-st_as_sfc(bb_buf)
bb_buf<-st_as_sf(bb_buf)

study_area<-bb_buf

# save as spatial data frame
save(study_area, file="0_data/manual/iButton_data/spatial/study_area.rData")

# save as shapefile
st_write(study_area, "0_data/manual/iButton_data/spatial/study_area.shp", delete_dsn = TRUE)

```

```{r eval=FALSE}
# Plot

# get basemap
# base<-basemap_raster(study_area, map_service = "esri", map_type = "delorme_world_base_map")
# base<-basemap_raster(study_area, map_service = "osm", map_type = "topographic")
# base<-basemap_raster(study_area, map_service = "osm_stamen", map_type = "terrain_bg")
# base<-basemap_raster(study_area, map_service = "esri", map_type = "world_shaded_relief")
base<-basemap_raster(study_area, map_service = "esri", map_type = "world_physical_map")

#alberta boundary
alberta<-st_read("0_data/external/Alberta/Alberta.shp")%>%st_transform(crs=st_crs(RIVR_xy))



# get aspect ratio of the study area (for inset map)
#asp <- (study_area$ymax - study_area$ymin)/(study_area$xmax - study_area$xmin)

mypal= c('#1b9e77','#d95f02','#7570b3','#e7298a')
# m<-tm_shape(alberta)+tm_borders()+tm_fill(col = "#fddadd")+
#   #tm_polygons(col=NA, border.col="black")+
#   tm_layout(frame=FALSE)+
#   tm_legend(outside=TRUE, frame=FALSE)+
m<-tm_shape(base)+
  tm_rgb()+
  tm_shape(ss_xy)+
    tm_symbols(col = "Project", palette = mypal, border.lwd = 0, size = .2, alpha=.8, legend.format = list(text.align="right"),
               legend.hist = TRUE)+
  tm_layout(title.size = 0.6,legend.outside = TRUE)+
tm_graticules(lines=FALSE)
  #tm_legend(position=c("left", "top"), frame=TRUE)


insetmap<-tm_shape(alberta)+tm_fill(col="lightgrey")+tm_borders(lwd=.9, col="black")+
  tmap_options(check.and.fix = TRUE)+
  tm_shape(study_area)+tm_borders(lw=2, col="red") +
  tm_layout(inner.margins = c(0.04,0.04,0.04,0.04), outer.margins=c(0,0,0,0), bg.color="transparent", frame = FALSE)
    #tm_symbols(shape = 20, alpha = .5, border.col="dimgray", size = .1, col = "black")
  #tm_legend(position=c("left", "top"), frame=TRUE)

#Get aspect ratio of bounding box
study_area_2<-st_bbox(study_area)
asp <- (study_area_2$ymax - study_area_2$ymin)/(study_area_2$xmax - study_area_2$xmin)

library(grid)
w <- .3
h <- asp * w
vp <- viewport(0.76, 0.23, width = w, height=h)
#vp <- viewport(0.9, 0.22, width = w, height=h, just=c("right", "top"))

m

tmap_save(m, "3_output/maps/ss_xy.png",  dpi=300, insets_tm=insetmap, insets_vp=vp,
          height=150, width=150, units="mm")

```

```{r studyArea, echo=FALSE, fig.cap= "iButton locations.", out.width = "50%"}
knitr::include_graphics("3_output/maps/ss_xy.png")
```
