==> 1_code/r_notebooks/3_ibutton_qualityControl.Rmd <==
```{r setup, include=FALSE, cache=FALSE}
#Set root directory to R project root
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

setwd(rprojroot::find_rstudio_root_file())
```

```{r eval=FALSE}

library(tidyr)
library(plyr)
library(dplyr)
library(kableExtra)
library(lubridate)
library(imputeTS)
library(readr)
library(stringr)
library(ggplot2)

```

```{r eval=FALSE}
# load ibutton daily summaries
load("2_pipeline/tmp/ibuttons.rData")
```

### Remove months with incomplete data {.unnumbered}

Removing first month of data if it has only 20 days or less.

```{r eval=FALSE}
ibuttons_1<- ibuttons%>% ungroup()%>%split(f=.$Site_StationKey) %>% lapply(FUN=function(x){
  if (x %>% filter(month(Date)==month(min(Date))& year(Date)==min(year(Date))) %>%
      summarize(max(Day)-min(Day)) %>% c () < 20 ){
    y<-x %>% filter(month(Date)>month(min(Date))| year(Date)> min(year(Date)))    
    
    return(y)
  }else{
    return(x)
  }
}) 
```

Removing last month of data if it has only 20 days or less.

```{r eval=FALSE}
ibuttons_2<-ibuttons_1 %>% lapply(FUN=function(x){
  if (x %>% filter(month(Date)==month(max(Date))& year(Date)==max(year(Date))) %>% 
      summarize(max(Day)-min(Day)) %>% c () < 20 ){
    y<-x %>% filter(month(Date)<month(max(Date))| year(Date)< max(year(Date)))    
    #y<-x[!(month(x$Date)==month(min(x$Date) & year(x$Date)==min(year(x$Date)))),]
    return(y)
  }else{
    return(x)
  }
  
}) %>% do.call(rbind,.) #working
```

### Calculate the difference bewteen ERA5 and iButton temps {.unnumbered}

Import ERA 5 daily summaries calculated in google earth engine.

```{r eval=FALSE}
ERA5_d<-read_csv("0_data/manual/gee_tables/ibutton_timeSeries_ERA5_daily.csv")

# convert kelvin to celsius

ERA5_d2<-ERA5_d%>%
  mutate(ERA5_temp_mean= mean_2m_air_temperature-273.15)%>%
  mutate(ERA5_temp_max= maximum_2m_air_temperature-273.15)%>%
  mutate(ERA5_temp_min= minimum_2m_air_temperature-273.15)%>%
  dplyr::rename(c(Year=year, Month=month, Day=day, Site_StationKey=St_SttK))%>%
  select(Site_StationKey, Year, Month, Day, ERA5_temp_max, ERA5_temp_min, ERA5_temp_mean)
```

Calculate difference between ibutton and ERA5 temperatures

```{r eval=FALSE}
ibuttons_ERA5<-ibuttons_2%>%
  left_join(ERA5_d2, by = c("Site_StationKey", "Day", "Month", "Year"))%>%
  mutate(Tmax_dif=ERA5_temp_max-Tmax_Day)%>%
  mutate(Tavg_dif=ERA5_temp_mean-Tavg_Day)%>%
  mutate(Tmin_dif=ERA5_temp_min-Tmin_Day)%>%
  filter(is.na(iBt_type)|iBt_type!="EXTRA")%>%
  # add season columns
  mutate(season_4 = case_when(
      Month %in%  9:11 ~ "Fall",
      Month %in%  c(12, 1, 2)  ~ "Winter",
      Month %in%  3:5  ~ "Spring",
      TRUE ~ "Summer"))%>%
  mutate( season_2 = case_when(
      Month %in%  c(10:12, 1, 2,3) ~ "Winter",
      TRUE ~ "Summer"))
```

### Remove outliers {.unnumbered}

```{r eval=FALSE}

#remove outliers 

ibuttons_ERA5_2<-ibuttons_ERA5%>%
# remove extreme values
    filter(Tmax_Day< 50)%>%
# remove outliers based on sd
    group_by(Site_StationKey, season_2) %>%
    filter(between(Tmax_Day, mean(Tmax_Day)-(3*sd(Tmax_Day)), mean(Tmax_Day)+(3*sd(Tmax_Day))))%>%
    filter(between(Tmin_Day, mean(Tmin_Day)-(3*sd(Tmin_Day)), mean(Tmin_Day)+(3*sd(Tmin_Day))))%>%
    filter(between(Tavg_Day, mean(Tavg_Day)-(3*sd(Tavg_Day)), mean(Tavg_Day)+(3*sd(Tavg_Day))))%>%
    ungroup()%>%
    filter(is.na(Tmax_dif)|between(Tmax_dif, mean(Tmax_dif,  na.rm=TRUE)-(3*sd(Tmax_dif, na.rm=TRUE)), mean(Tmax_dif,  na.rm=TRUE)+(3*sd(Tmax_dif, na.rm=TRUE))))%>%
filter(is.na(Tmin_dif)|between(Tmin_dif, mean(Tmin_dif,  na.rm=TRUE)-(3*sd(Tmin_dif, na.rm=TRUE)), mean(Tmin_dif,  na.rm=TRUE)+(3*sd(Tmin_dif, na.rm=TRUE))))%>%
filter(is.na(Tavg_dif)|between(Tavg_dif, mean(Tavg_dif,  na.rm=TRUE)-(3*sd(Tavg_dif, na.rm=TRUE)), mean(Tavg_dif,  na.rm=TRUE)+(3*sd(Tavg_dif, na.rm=TRUE))))


## remove outliers based on interquartile range
# 
# d1<-d %>%
#     group_by(Site_StationKey, season_2) %>%
#     filter(between(Tmax_Day, quantile(Tmax_Day, .25, )-(3*IQR(Tmax_Day)), quantile(Tmax_Day, .75, )+(3*IQR(Tmax_Day))))%>%
#     filter(between(Tmin_Day, quantile(Tmin_Day, .25, )-(3*IQR(Tmin_Day)), quantile(Tmin_Day, .75, )+(3*IQR(Tmin_Day))))%>%
#     filter(between(Tavg_Day, quantile(Tavg_Day, .25, )-(3*IQR(Tavg_Day)), quantile(Tavg_Day, .75, )+(3*IQR(Tavg_Day))))%>%
#     ungroup()%>%
#     filter(is.na(Tmax_dif)|between(Tmax_dif, quantile(Tmax_dif, .25, na.rm=TRUE)-(3*IQR(Tmax_dif, na.rm=TRUE)), quantile(Tmax_dif, .75, na.rm=TRUE)+(3*IQR(Tmax_dif, na.rm=TRUE))))%>%
# filter(is.na(Tmin_dif)|between(Tmin_dif, quantile(Tmin_dif, .25, na.rm=TRUE)-(3*IQR(Tmin_dif, na.rm=TRUE)), quantile(Tmin_dif, .75, na.rm=TRUE)+(3*IQR(Tmin_dif, na.rm=TRUE))))%>%
# filter(is.na(Tavg_dif)|between(Tavg_dif, quantile(Tavg_dif, .25, na.rm=TRUE)-(3*IQR(Tavg_dif, na.rm=TRUE)), quantile(Tavg_dif, .75, na.rm=TRUE)+(3*IQR(Tavg_dif, na.rm=TRUE))))

  
save(ibuttons_ERA5_2, file="2_pipeline/tmp/ibuttons_ERA5_2.rData")
```

### Impute missing data {.unnumbered}

#### Create a dummy iButton data frame

The data frame will have rows for every day during the time period the iButtons were deployed. Daily temperature columns will be filled with NA values.

Create a calendar for the time iButtons were deployed.

```{r eval=FALSE}
ibuttons_ERA5_3<-ibuttons_ERA5_2%>%
  mutate(project_year=str_c(Project, "_", Year))%>%
  mutate(project_year_month=str_c(Project, "_", Year, "_", Month))%>%
  select("Project","Site_StationKey", "Day", "Month","Year", "project_year", "project_year_month", "Date", "iBt_type",  "Tmax_Day","Tmin_Day",  "Tavg_Day")

months <- 1:12

# create a data frame with the number of days per month
days <- days_in_month(months) %>% as.data.frame() %>% `colnames<-`("Days") %>%
  mutate(Month=row.names(.))

# create a list of months 
calendar<-apply(days,MARGIN = 1,FUN=function(x){
 seq(1:x)
  } 
 ) %>%
  lapply(.,FUN = unlist)

#create a calendar data frame with all of the days of the year for all stations. The dataframe will have columns for day, month,  and year and spans the time of the ibuttons
calendar_1<-calendar %>% `names<-`(days$Month) %>% unlist() %>% as.data.frame() %>%
  `colnames<-`("Day") %>% mutate(Month_name=substr(rownames(.),1,3)) %>%
  left_join(data.frame(Month_name=month.abb,Month=months)) %>%
  expand_grid(.,Year=c(min(ibuttons_ERA5_3$Year):max(ibuttons_ERA5_3$Year)))%>%
  mutate(Date = dmy(paste(Day, Month, Year, sep = "/")))
```

Create the dummy data frame that includes all of the days/year of the above calendar data frame with NA instead of temperature values.

```{r eval=FALSE}

calendar_2<-ibuttons_ERA5_3%>%
  ungroup%>%
  select(Site_StationKey, Project)%>%
  unique() %>%
 `colnames<-` (c("Site_StationKey","Project")) %>%
  expand_grid(.,calendar_1)%>% mutate(iBt_type=NA,Tmax_Day=NA,Tmin_Day=NA,Tavg_Day=NA)%>%
  mutate(project_year_month=str_c(Project, "_", Year, "_", Month))%>%
  filter(project_year_month%in% unique(ibuttons_ERA5_3$project_year_month))%>%
  select(-c(iBt_type))%>%
  arrange(Site_StationKey, Year, Month, project_year_month, Project)
```

Add days with missing data to ibutton dataframe

```{r eval=FALSE}

# add days with missing data to ibutton dataframe
complete_data_w_missing<-calendar_2 %>% ungroup() %>% 
  select(-c(Tmax_Day, Tmin_Day, Tavg_Day)) %>%
  left_join(ibuttons_ERA5_3)%>%
  distinct()
```

#### Impute missing values {.unnumbered}

Impute NA values using a spline function based on time series imputation. The imputation is based on month per year per iButton site.

```{r eval=FALSE}


ibuttons_complete_daily<-complete_data_w_missing %>%
  arrange(Site_StationKey, Date)%>%
  select(-Month_name)%>%
  # group_by(Site_StationKey,Year, Month) %>%
  group_by(Site_StationKey) %>%
  # group_modify(~ na.trim(.)) %>%  remove the NA's at the beginning and end of each group.
  mutate(Tmax_Day_int = na_interpolation(Tmax_Day, option="linear", maxgap = 10))%>%
  mutate(Tmin_Day_int = na_interpolation(Tmin_Day, option="linear", maxgap = 10))%>%
  mutate(Tavg_Day_int = na_interpolation(Tavg_Day, option="linear", maxgap = 10))%>%
  mutate(imputed= ifelse(is.na(Tmax_Day), "yes", "no"))%>%
  select(-c(Tmax_Day, Tmin_Day, Tavg_Day))%>%
  rename(Tmax_Day =Tmax_Day_int, Tmin_Day =Tmin_Day_int, Tavg_Day =Tavg_Day_int,)

  
#save
save(ibuttons_complete_daily, file="0_data/manual/iButton_data/ibuttons_complete_daily.rData")

write.csv(ibuttons_complete_daily, file="0_data/manual/iButton_data/ibuttons_complete_daily.csv")

```

### Compare imputed values with ERA5 dailys {.unnumbered}

```{r eval=FALSE}

ibuttons_ERA5_4<-ibuttons_complete_daily%>%
  left_join(ERA5_d2, by = c("Site_StationKey", "Day", "Month", "Year"))%>%
  mutate(Tmax_dif=ERA5_temp_max-Tmax_Day)%>%
  mutate(Tavg_dif=ERA5_temp_mean-Tavg_Day)%>%
  mutate(Tmin_dif=ERA5_temp_min-Tmin_Day)%>%
  mutate(T_diurnal=Tmax_Day-Tmin_Day)%>%
  mutate(Date = make_date(Year, Month, Day))%>%
  mutate(season_4 = case_when(
      Month %in%  9:11 ~ "Fall",
      Month %in%  c(12, 1, 2)  ~ "Winter",
      Month %in%  3:5  ~ "Spring",
      TRUE ~ "Summer"))%>%
  mutate( season_2 = case_when(
      Month %in%  c(10:12, 1, 2,3) ~ "Winter",
      TRUE ~ "Summer"))
```


### Monthly summaries {.unnumbered}

#### Snow burial {.unnumbered}

We defined snow burial as ibuttons with a diural range of <3 degrees for 25 consequetive days or more. [@wood2017dtdf]

```{r eval=FALSE}
#define how may consecutive days are needed to define as snow
snow_thresh<-25

# Flag snow burials
ibuttons_ERA5_5<-ibuttons_ERA5_4%>%
  filter(T_diurnal<3)%>% #filter to diurnal range <3
  # mutate(lag_date=Date - days(1))%>%
  group_by(Site_StationKey)%>%
  arrange(Site_StationKey, Date)%>%
  # mutate(consecutive_day_flag = if_else(Date == (Date - days(1), 1, 0))%>%
  mutate(lag_date=lag(Date))%>%
  mutate(consecutive_day_flag = if_else(Date == (lag(Date) + days(1)), 1, 0))%>% # flag consecutive days
  mutate(consecutive_day_flag =replace_na(consecutive_day_flag, 0))%>%# flag consecutive days
  mutate(snow_burial=rep(rle(consecutive_day_flag)$length>=snow_thresh,rle(consecutive_day_flag)$length))%>% #flag dates that are a part of  sequence
  right_join(ibuttons_ERA5_4)%>% # join back with the original data frame
  mutate(snow_burial=replace_na(snow_burial, "FALSE"))%>%
  select(-c(consecutive_day_flag, lag_date))%>%
  arrange(Site_StationKey, Date)%>%
# Replace flaged snow burials with NA  
  mutate(Tmax_Day= ifelse(snow_burial=="TRUE", NA, Tmax_Day))%>%
  mutate(Tmax_Day= ifelse(snow_burial=="TRUE", NA, Tmin_Day))%>%
  mutate(Tmax_Day= ifelse(snow_burial=="TRUE", NA, Tavg_Day))

```

#### Remove months with more than 10 missing days {-}
```{r eval=FALSE}
ibuttons_ERA5_6<-ibuttons_ERA5_5%>% 
  group_by(Project, Site_StationKey,Month,Year) %>%
  dplyr::summarize(count_na = sum(is.na(Tmax_Day)))%>%
  right_join(ibuttons_ERA5_5 ,by=c("Project", "Site_StationKey","Month","Year"))%>%
  filter(count_na<10)


save(ibuttons_ERA5_6, file="2_pipeline/tmp/ibuttons_ERA5_6.rData")
```



#### Calculate monthly summaries {-}

```{r eval=FALSE}
load("2_pipeline/tmp/ibuttons_ERA5_6.rData")


ibuttons_complete_monthly<-ibuttons_ERA5_6%>%
        group_by(Project, Site_StationKey,Month,Year)%>%
        summarise(across(c(Tmax_Day, Tmin_Day, Tavg_Day), mean, na.rm=T))%>%
        rename(c(Tmax_Month=Tmax_Day, Tmin_Month=Tmin_Day, Tavg_Month=Tavg_Day))%>%
        arrange(Site_StationKey, Year, Month)

save(ibuttons_complete_monthly, file="0_data/manual/iButton_data/ibuttons_complete_monthly.rData")

write.csv(ibuttons_complete_monthly, file="0_data/manual/iButton_data/ibuttons_complete_monthly.csv")

```

```{r echo=FALSE, message = FALSE, results="asis"}
load("0_data/manual/iButton_data/ibuttons_complete_monthly.rData")

knitr::kable(head(ibuttons_complete_monthly), "pipe") 
```

### Difference between monthly summaries {.unnumbered}

```{r eval=FALSE}
ERA5_m<-read_csv("0_data/manual/gee_tables/ibutton_timeSeries_ERA5_monthly.csv")

 # convert kelvin to celsius

ERA5_m2<-ERA5_m%>%
  mutate(ERA5_temp_mean= mean_2m_air_temperature-273.15)%>%
  mutate(ERA5_temp_max= maximum_2m_air_temperature-273.15)%>%
  mutate(ERA5_temp_min= minimum_2m_air_temperature-273.15)%>%
  rename(c(Year=year, Month=month, Site_StationKey=St_SttK))%>%
  select(Site_StationKey, Year, Month, ERA5_temp_max, ERA5_temp_min, ERA5_temp_mean)


ibutton_timeSeries_ERA5_monthly_2<-ibuttons_complete_monthly%>%
  left_join(ERA5_m2, by = c("Site_StationKey", "Month", "Year"))%>%
  mutate(Tmax_dif_m=ERA5_temp_max-Tmax_Month)%>%
  mutate(Tavg_dif_m=ERA5_temp_mean-Tavg_Month)%>%
  mutate(Tmin_dif_m=ERA5_temp_min-Tmin_Month)%>%
  # filter(is.na(iBt_type)|iBt_type!="EXTRA")%>%
  # add season columns
  mutate(season_4 = case_when(
      Month %in%  9:11 ~ "Fall",
      Month %in%  c(12, 1, 2)  ~ "Winter",
      Month %in%  3:5  ~ "Spring",
      TRUE ~ "Summer"))%>%
  mutate( season_2 = case_when(
      Month %in%  c(10:12, 1, 2,3) ~ "Winter",
      TRUE ~ "Summer"))

hist(ibutton_timeSeries_ERA5_monthly_2$Tavg_dif_m)
hist(ibutton_timeSeries_ERA5_monthly_2$Tmin_Month)
hist(ibutton_timeSeries_ERA5_monthly_2$ERA5_temp_min)
```
