==> 1_code/r_notebooks/4_ibutton_climateNA.Rmd <==
```{r setup, include=FALSE, cache=FALSE}
#Set root directory to R project root
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

## Setup {.unnumbered}

**Install `ClimateNAr`**

Get the latest ClimateBC/NA r package by registering at a <https://register.climatena.ca/>. Follow the instructions to install the `ClimateNAr` package.
 
```{r eval=FALSE}
library(ClimateNAr)
library(dplyr)
library(sf)
library(tidyr)
library(stringr)
```

Download the ClimateNA desktop application and r package here: https://register.climatena.ca/

Instructions on how use the ClimateNA application can be found here: https://pressbooks.bccampus.ca/fode010notebook/chapter/topic-3-2-the-use-of-climatena-ap-to-generate-point-and-spatial-climate-data/


The climateNA desktop application was designed for PC, but can be used on Mac by using Wine. To get climateNA working on my machine I used a Wineskin (https://github.com/Gcenx/WineskinServer) to create an application wrapper. The wrapper used the WS10Wine64Bit5.20 wine engine. I also installed `vb6run` with winetricks to get the application working. The the wine wrapper should contain the entire `climateNA_v730` folder directory, not just the `ClimateNA_v7.30.exe` file. 

While climateNA can be run via commandline using R, I was not able to get this working on my Mac. I included code for integrating climateNA into the R workflow on PC's ([Extract CLimateNA monthly summaries with R ]). However, the R code is untested. 


## Create input file {.unnumbered}

ClimateNAr requires a properly formatted .csv input file that includes the following headers: ID1, ID2, lat, long, el


1.  Bring in iButton data\*\*

```{r eval=FALSE}
# load("0_data/manual/iButton_data/ibuttons_complete_monthly.rData")
load("0_data/manual/iButton_data/spatial/ss_xy.rData")

# add month and year columns to the ibutton spatial dataframe
# ibutton<-ss_xy%>%
#       left_join(ibuttons_complete_monthly)%>%
#       select(Project, Site_StationKey, Month, Year)
```

2.  Bring in the elevation data extracted via Google Earth Engine.

```{r eval=FALSE}
elev<-read.csv("0_data/manual/gee_tables/ibutton_terrain.csv")
elev$Project[elev$Project== "WOOD"] <- "wood"
elev$Project[elev$Project== "HL"] <- "HILLS"

```

3.  Join elevation data to the iButton spatial data frame.

```{r eval=FALSE}
load("0_data/manual/iButton_data/spatial/ss_xy.rData")

climateNA_input<- ss_xy%>%
    left_join(elev, by=c('Project'='Project', 'Site_StationKey'='St_SttK'))%>%
    as.data.frame()%>%
    select(Project, Site_StationKey, Lat, Long, elevation)%>%
    rename(c(ID1=Project, ID2=Site_StationKey, lat=Lat, long=Long, el=elevation))%>%
    drop_na()

write.csv(climateNA_input, file="0_data/manual/climateNA/input/climateNA_input.csv", row.names = FALSE)  
```

## Extract ClimateNA summaries {-}

I used the ClimateNA GUI to extract monthly climate summaries. There is a weird bug on my mac ClimateNA application where I need to open and save the input .csv in excel before ClimateNA can read it.   

In the desktop application I selected a `Historical Time Series` of `Monthly primary variables (60)`

Climate data is exported back into the rProject as a .csv file in `0_data/manual/climateNA/output/`

### Read in the ClimateNA summaries {-}

```{r eval =FALSE}
cna<-read.csv(file="0_data/manual/climateNA/output/climateNA_input_2005-2021MP.csv")
  
cNA_month_summaries<-cna%>%
  rename(c(Project=ID1, Site_StationKey=ID2))%>%
  dplyr::select(c(Year:Tave12))%>%
  gather(key="Month_a", value="Tmax_cNA", "Tmax01":"Tmax12")%>%
  mutate(Month_a=(gsub("[^0-9.-]", "", Month_a))) %>%
  gather(key="Month_b", value="Tmin_cNA", "Tmin01":"Tmin12")%>%
  mutate(Month_b=(gsub("[^0-9.-]", "", Month_b))) %>%
  gather(key="Month_c", value="Tavg_cNA", "Tave01":"Tave12")%>%
  mutate(Month_c=(gsub("[^0-9.-]", "", Month_c))) %>%
  filter(Month_a==Month_b & Month_b==Month_c)%>%
  dplyr::select(Project, Site_StationKey, Year, Month_a, Tmax_cNA, Tmin_cNA, Tavg_cNA)%>%
  rename(Month=Month_a)%>%
  mutate(Month=as.numeric(str_remove(Month, "^0+")))%>%
  arrange(Site_StationKey, Year, Month)

save(cNA_month_summaries, file="0_data/manual/climateNA/processed/cNA_month_summaries.rData")
write.csv(cNA_month_summaries, file="0_data/manual/climateNA/processed/cNA_month_summaries.csv")
```

```{r}
load("0_data/manual/climateNA/processed/cNA_month_summaries.rData")
knitr::kable(head(cNA_month_summaries), "pipe") 
```



### Extract ClimateNA monthly summaries directly with R {-}

I have not been able to get this working on my Mac, but it should work on PC's (untested).


```{r eval=FALSE}
library(ClimateNAr)
#Inputs
#ClimateNA executable
c<-"ClimateNA_v7.30.exe"
#input directory
#e.g."/C:\\Users\\kimorris\\ClimateNA_v730\\ibutton\\"
#data<-'/C:\\Users\\kimorris\\ClimateNA_v730\\ibutton\'
#Automated
maxYear<-max(tmax2$year)
minYear<-min(tmax2$year)
for (y in seq(minYear,maxYear)){
  #loop through for each year
  setwd(climNA_dir);getwd() # it must be the home directory of ClimateNA
  exe <- c
  #inputFile = paste0(data,"temp_input.csv") #find way to change exported file location earlier into format needed here?
  #outputFile = paste0(data,"temp_input_Year_",y,"MP.csv")
  inputFile = paste0("/C:\\Users\\kimorris\\ClimateNA_v730\\ibutton\temp_input.csv") #find way to change exported file location earlier into format needed here?
  outputFile = paste0("/C:\\Users\\kimorris\\ClimateNA_v730\\ibutton\temp_input.csv","temp_input_Year_",y,"MP.csv")
  yearPeriod = paste0('/Year_',y,'.ann')
  system2(exe,args= c('/Y', yearPeriod, inputFile, outputFile)) 
  #dat <- read.csv('C:/Users/kimorris/ClimateNA_v730/ibutton/temp_output.csv'); head(dat) 
}

maxYear<-max(tmax2$year)
minYear<-min(tmax2$year)
ib_clim<-NULL
for (y in seq(minYear,maxYear)){
  file<-paste0(working_dir,"temp_input_Year_",y,"MP.csv")
  print(file)
  t<-read.csv(file)
  print(head(t))
  tt<-subset(t,select=c("ID1","ID2","Latitude","Longitude","Elevation", "Tmax01","Tmax02","Tmax03","Tmax04","Tmax05","Tmax06","Tmax07","Tmax08","Tmax09","Tmax10","Tmax11","Tmax12"))
  print(head(tt))
  t_long <- melt(setDT(tt), id.vars = c("ID1","ID2","Latitude","Longitude","Elevation"), variable.name = "month")
  print(head(t_long))
  #select only needed columns
  t_long2<-subset(t_long,select=c("ID1","month","value"))
  #parse month from Tmax01 etc
  t_long2$month<-str_sub(t_long2$month, start= -2)
  #name value column ClimNATmax
  colnames(t_long2)<-c("group","month","climNA_tmax")
  #remove leading 0's on month
  t_long2$month<-sub("^0+", "",t_long2$month)
  #head(t_long2)
  #table(t_long2$month)
  #merge with tmax2
  t_tmax<-subset(tmax2,year==y)
  #table(t_tmax$month)
  t_tmax2<-merge(t_tmax,t_long2,by=c("group","month"))
  #head(t_tmax2)
  #table(t_tmax2$month)
  #nrow(t_tmax2)
  #nrow(t_tmax)
  ##merge years as it loops through them
  ib_clim<-rbind(ib_clim,t_tmax2)
}
```

## Calculate difference between iButton and ClimateNA summaries {-}

```{r eval=FALSE}
# ClimateNA monthly summaries
load("0_data/manual/climateNA/processed/cNA_month_summaries.rData")

# ibBtton monthly summaries
load("0_data/manual/iButton_data/ibuttons_complete_monthly.rData")

iButton_cNA_diff<-ibuttons_complete_monthly%>%
  left_join(cNA_month_summaries, by=c("Project", "Site_StationKey", "Year", "Month"))%>%
  mutate(Tmax_diff=round(Tmax_cNA-Tmax_Month, 2))%>%
  mutate(Tmin_diff=round(Tmin_cNA-Tmin_Month, 2))%>%
  mutate(Tavg_diff=round(Tavg_cNA-Tavg_Month, 2))%>%
  na.omit

# Plot 
iButton_cNA_diff%>%
  ggplot(aes(x=Tavg_cNA, y=Tavg_Month)) +
  geom_point(alpha=.5)

iButton_cNA_diff%>%toronto
  ggplot(aes(x=Tavg_diff)) +
  geom_histogram(aes(y=..density..), colour="black", fill="white", binwidth = .5)+
   geom_density(alpha=.2, fill="#FF6666") 

# over time
iButton_cNA_diff %>%
    mutate(date = lubridate::ymd(paste(Year, Month, 1)))%>%
  filter(Tavg_diff>-5)%>%
  ggplot(aes(date, Tmax_Month, group=Site_StationKey)) +
  geom_line(alpha=0.2) +
  scale_x_date(date_labels = "%b %Y", date_breaks = "1 month")+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 90, vjust=0.5), 
        panel.grid.minor = element_blank())

# Save 
# iButton_cNA_diff<-iButton_cNA_diff%>%
#   dplyr::select("Project","Site_StationKey","Month","Year", "Tmax_diff", "Tmin_diff", "Tavg_diff" )

save(iButton_cNA_diff, file="0_data/manual/iButton_data/iButton_cNA_diff.rData")
write.csv(iButton_cNA_diff, file="0_data/manual/iButton_data/iButton_cNA_diff.csv")

```
