==> 1_code/r_notebooks/5_covariates_gee_spatial.Rmd <==
```{r setup, include=FALSE, cache=FALSE}
#Set root directory to R project root
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

setwd(rprojroot::find_rstudio_root_file())

```

```{r eval=FALSE}
library(dplyr)
library(readr)
```

## GEE scripts {.unnumbered}

Bring GEE scripts into the current rProject by cloning the GEE git and copying over the .js files.

```{bash engine.opts='-l', eval=FALSE}
# set directory to clone GEE git into (should be empty)
cd 1_code/GEE

# delete all .js files in directory
rm *.js

# clone gee scripts
git clone https://earthengine.googlesource.com/users/bgcasey/climate_downscaling


# add the .js file extention to files
find climate_downscaling -type f -not -name "*.*" -exec mv "{}" "{}".js \;

# move .js files up a directory
mv -v climate_downscaling/*.js .

# move functions folder up a directory
mv -v climate_downscaling/functions/ .

# delete .git folder
rm -r climate_downscaling/
```

## GEE CSVs {.unnumbered}

```{r eval=FALSE}
gee_cov_all<-read_csv("0_data/manual/gee_tables/ibutton_all_indices.csv")
gee_cov_all$Project[gee_cov_all$Project== "WOOD"] <- "wood"
gee_cov_all$Project[gee_cov_all$Project== "HL"] <- "HILLS"
gee_cov_all<-rename(gee_cov_all, c("Site_StationKey" = "St_SttK", "Month"="month", "Year"= "year", "tree_coverfraction"="tree-coverfraction"))

### add latitude
load("0_data/manual/iButton_data/spatial/ss_xy.rData")

library(sf)

gee_cov_all<-dplyr::left_join(gee_cov_all, ss_xy, by = c("Project", "Site_StationKey"))%>%
  dplyr::select(-c(Long, min_year, max_year, geometry))

save(gee_cov_all, file="0_data/manual/gee_tables/gee_cov_all.rData")
```

### Join with response variable {.unnumbered}

```{r eval=FALSE}
load("0_data/manual/iButton_data/iButton_cNA_diff.rData")

data_full<-left_join(iButton_cNA_diff, gee_cov_all, by=c("Project", "Site_StationKey", "Month", "Year"))%>%
  dplyr::select(-date)%>%
  as.data.frame()
```

### Add season column {.unnumbered}

```{r eval=FALSE}

data_full<-data_full%>%
 mutate(season_4 = case_when(
      Month %in%  9:11 ~ "Fall",
      Month %in%  c(12, 1, 2)  ~ "Winter",
      Month %in%  3:5  ~ "Spring",
      TRUE ~ "Summer"))%>%
  mutate(season_2 = case_when(
      Month %in%  c(10:12, 1, 2,3) ~ "Winter",
      TRUE ~ "Summer"))

save(data_full, file="0_data/manual/formatted_for_models/data_full.rData")

write.csv(data_full, file="0_data/manual/formatted_for_models/data_full.csv", row.names = FALSE)  
```




## GEE Rasters

```{r eval=FALSE}
library(raster)
library(terra)
library(dplyr)
library(foreach)
library(stringr)
library(sf)
```

**Import and process GEE rasters**

The following code chunk:

1.   Imports raster tiles generated using Google Earth Engine.
2.   Stitches them together into a single continuous raster using the `raster::mosaic` function.
3.   Crops the continuous raster to the study area.
4.  Saves the raster as a .tif file.

```{r eval=FALSE}

# Load study area
study_area<-st_read("../Data/Canada/alberta_bc")

# dataPath <- "/Volumes/GoogleDrive/My Drive/PhD/thesis/chapter_1/empirical/0_data/external/CL_LiDAR/grid_metrics_surfaces/"
path_name<-"0_data/external/GEE_rasters/to_process/"

bl<-list.dirs(path=path_name, full.names = F, recursive = F)

#set temp directory to an external drive to free up harddrive space
rasterOptions(tmpdir=file.path("../r_tmp")) 

foreach (j = 1:length(bl)) %do%
{
  t1 = proc.time()
  print(paste("---- Begin",bl[j], Sys.time()))
  #list .tif files in the folder
  fl <- list.files(path=paste0(path_name, bl[j]), pattern = '*.tif$', recursive = T, full.names = T)
  print(paste("----",bl[j]," file scan complete", Sys.time()))
  # If multibabnd rasters, read each file into R as a raster brick. Use raster function if   single band rasters
  fl <- lapply(fl, brick)
  print(paste("----",bl[j]," lapply raster::brick  complete", Sys.time()))
  # set mosaic parameters
  fl$fun <- mean
  fl$tolerance <-.5
  x <- do.call(raster::mosaic, fl)
  # rename the raster bands. For multiband rasters, the band names are lost during mosaicing
  names(x)<-names(fl[[]]) # name mosaic bands
  names(x)<-gsub("_mean", "",names(x))
  print(paste("----",bl[j]," mosaic complete!", Sys.time()))
  x<-raster::crop(x, study_area) # crop raster to study area
  print(paste("----",bl[j]," crop complete", Sys.time()))
  x<-terra::rast(x) # preserve band names when savinbg by converting the stack to a rast object from the terra package.
  writeRaster(x, filename=paste0("0_data/manual/raster_mosaics/",bl[j], ".tif"),  overwrite=T)
  print(paste("----",bl[j]," saved as TIFF", Sys.time()))
  print("process time")
  print(proc.time() - t1)
  removeTmpFiles(0) #The raster package can store a lot of files. This removes any temp raster files generated   during the loop
}



#######################################


# foreach (j = 1:length(bl)) %do%
# {
#   t1 = proc.time()
#   print(paste("---- Begin",bl[j], Sys.time()))
#   #list .tif files in the folder
  fl <- list.files(path=paste0(path_name, bl[2]), pattern = '*.tif$', recursive = T, full.names = T)
  print(paste("----",bl[j]," file scan complete", Sys.time()))
  # If multibabnd rasters, read each file into R as a raster brick. Use raster function if   single band rasters
  fl <- lapply(fl, brick)
  print(paste("----",bl[j]," lapply raster::brick  complete", Sys.time()))
  # set mosaic parameters
  fl$fun <- mean
  fl$tolerance <-.5
  x <- do.call(raster::mosaic, fl)
  # rename the raster bands. For multiband rasters, the band names are lost during mosaicing
  names(x)<-names(fl[[]]) # name mosaic bands
  names(x)<-gsub("_mean", "",names(x))
  print(paste("----",bl[j]," mosaic complete!", Sys.time()))
  x<-raster::crop(x, study_area) # crop raster to study area
  print(paste("----",bl[j]," crop complete", Sys.time())) 
  x<-terra::rast(x) # preserve band names when savinbg by converting the stack to a rast object from the terra package.
  writeRaster(x, filename=paste0("0_data/manual/raster_mosaics/",bl[j], ".tif"),  overwrite=T)
  print(paste("----",bl[j]," saved as TIFF", Sys.time()))
  print("process time")
  print(proc.time() - t1)
  removeTmpFiles(0) #The raster package can store a lot of files. This removes any temp raster files generated   during the loop
}


```

## Setup data for GBM {-}

Set up a data frame to be used by the `gbm` function. It will only have
the predictor and response variables.

```{r eval= FALSE}
load("0_data/manual/formatted_for_models/data_full.rData")

df<-data_full%>%
  mutate(season_2=as.factor(season_2))%>%
  mutate(season_4=as.factor(season_4))

df1<-df%>%
dplyr::select(Tmax_diff, Tmin_diff, Tavg_diff, NDVI, NDMI, LAI, TWI, elevation, slope, HLI, tpi_50, tpi_500, northness, snow, NDSI, pr, srad, vs, soil, canopy_height, canopy_standard_deviation, tree_coverfraction, cloud_fraction, discrete_classification, forest_type, CHILI, Month, season_4)
```

### Check for multicollinearity between predictors {.unnumbered}

```{r eval=FALSE}
pairs_cov <-df1[c(4:23, 26)]

#visualize with corrplot. Easier to visualize with a lot of variables
M<-cor(pairs_cov, method = "pearson", use="pairwise.complete.obs")
corrplot(M, tl.cex=0.5, method="number", type ="upper", addCoefasPercent=TRUE, order = "hclust", number.cex=.5, cl.cex=.5
)

# remove highly correlated variables
#compare correlated predictors in univariate models
m1<-lm(Tavg_diff~snow, data=df1)
m2<-lm(Tavg_diff~NDSI, data=df1)
m3<-lm(Tavg_diff~CHILI, data=df1)
m4<-lm(Tavg_diff~HLI, data=df1)
m5<-lm(Tavg_diff~tree_coverfraction, data=df1)
m6<-lm(Tavg_diff~canopy_height, data=df1)

df2<-df1%>%
  dplyr::select(-c(NDSI, tree_coverfraction, HLI ))

data_sub<-df2
save(data_sub, file="0_data/manual/formatted_for_models/data_sub.rData")
```
