==> 1_code/r_notebooks/7_make_predictive_offset_raster.Rmd <==

```{r setup, include=FALSE, cache=FALSE}
#Set root directory to R project root
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```


```{r eval=FALSE}
library(dismo)
library(raster)
library(gbm)
```


### Load spatial variable rasters {.unnumbered}

```{r eval=FALSE}
ts_summer<-raster::brick("0_data/manual/raster_mosaics/time_series_06.tif")

# test<-brick("0_data/external/processed_GEE/time_series_06/neighborhood_rasters_4s_2021_6-0000065536-0000065536.tif")
# names(ts_summer)<-gsub("_mean", "",names(test))

ts_winter<-raster::brick("0_data/manual/raster_mosaics/time_series_12.tif")
ts_spring<-raster::brick("0_data/manual/raster_mosaics/time_series_03.tif")
ts_fall<-raster::brick("0_data/manual/raster_mosaics/time_series_09.tif")

fixed<-raster::brick("0_data/manual/raster_mosaics/fixed.tif")

# Load data
load("0_data/manual/formatted_for_models/data_full.rData")
df<-data_full%>%
  mutate(season_2=as.factor(season_2))%>%
  mutate(season_4=as.factor(season_4))
df5<-df
```

#### Combine raster mosaics {-}

```{r eval=FALSE}
gee_all_summer<-raster::stack(fixed, ts_summer)
crs(gee_all_summer)<-"EPSG:3348"

gee_all_spring<-stack(fixed, ts_spring)
crs(gee_all_spring)<-"EPSG:3348"

gee_all_winter<-stack(fixed, ts_winter)
crs(gee_all_winter)<-"EPSG:3348"

gee_all_fall<-stack(fixed, ts_fall)
crs(gee_all_fall)<-"EPSG:3348"


## Use smaller study area to test.
# test prediction code on a smaller raster area to reduce processing time

# load("../chapter_1/empirical/0_data/manual/bird/studyarea_big.rData")
# # study_area<-st_transform(c_bb, crs="EPSG:3348")
# # 
# # gee_all_summer_1<-crop(gee_all_summer, study_area)
```

### Difference between mean temperature {.unnumbered}

#### Summer {-}

```{r eval=FALSE}
load("2_pipeline/store/models/brt_meanTemp_summer.rData")

# There is no raster for season so we'll create a data frame with a constant value to plug into the predict function.
# season_4 <- factor('Summer', levels = levels(df5$season_4))
# add <- data.frame(season_4)
# add to predict function using cons=add
# p_summer <- predict(gee_all_summer, brt_meanTemp_summer, 
#        n.trees=brt_meanTemp_summer$gbm.call$best.trees, type="response")

p_summer <- predict(gee_all_summer, brt_meanTemp_summer,
       n.trees=brt_meanTemp_summer$gbm.call$best.trees, type="response")

names(p_summer)<-"Tmean_diff_summer"

writeRaster(p_summer, file="3_output/offset_rasters/meanTemp_summer_offset.tif", format='GTiff', overwrite=TRUE)
```

#### Fall {-}
```{r eval=FALSE}
load("2_pipeline/store/models/brt_meanTemp_fall.rData")

p_fall <- predict(gee_all_fall, brt_meanTemp_fall,
       n.trees=brt_meanTemp_fall$gbm.call$best.trees, type="response")

names(p_fall)<-"Tmean_diff_fall"

writeRaster(p_fall, file="3_output/offset_rasters/meanTemp_fall_offset.tif", format='GTiff', overwrite=TRUE)
```


#### Winter {-}

```{r eval=FALSE}
load("2_pipeline/store/models/brt_meanTemp_winter.rData")

p_winter <- predict(gee_all_winter, brt_meanTemp_winter, 
       n.trees=brt_meanTemp_winter$gbm.call$best.trees, type="response")

names(p_winter)<-"Tmean_diff_winter"

writeRaster(p_winter, file="3_output/offset_rasters/meanTemp_winter_offset.tif", format='GTiff', overwrite=TRUE)
```
 
#### Spring {-}

```{r eval=FALSE}
load("2_pipeline/store/models/brt_meanTemp_spring.rData")

p_spring <- predict(gee_all_spring, brt_meanTemp_spring, 
       n.trees=brt_meanTemp_spring$gbm.call$best.trees, type="response")

names(p_spring)<-"Tmean_diff_spring"

writeRaster(p_spring, file="3_output/offset_rasters/meanTemp_spring_offset.tif", format='GTiff', overwrite=TRUE)
```


### Difference between max temperature {.unnumbered}

#### Summer {-}

```{r eval=FALSE}
load("2_pipeline/store/models/brt_maxTemp_summer.rData")

p_summer <- predict(gee_all_summer, brt_maxTemp_summer, 
       n.trees=brt_maxTemp_summer$gbm.call$best.trees, type="response")

names(p_summer)<-"Tmax_diff_summer"

writeRaster(p_summer, file="3_output/offset_rasters/maxTemp_summer_offset.tif", format='GTiff', overwrite=TRUE)
```

#### Fall {-}
```{r eval=FALSE}
load("2_pipeline/store/models/brt_maxTemp_fall.rData")

p_fall <- predict(gee_all_fall, brt_maxTemp_fall, 
       n.trees=brt_maxTemp_fall$gbm.call$best.trees, type="response")

names(p_fall)<-"Tmax_diff_fall"

writeRaster(p_fall, file="3_output/offset_rasters/maxTemp_fall_offset.tif", format='GTiff', overwrite=TRUE)
```

#### Winter {-}

```{r eval=FALSE}
load("2_pipeline/store/models/brt_maxTemp_winter.rData")

p_winter <- predict(gee_all_winter, brt_maxTemp_winter, 
       n.trees=brt_maxTemp_winter$gbm.call$best.trees, type="response")

names(p_winter)<-"Tmax_diff_winter"

writeRaster(p_winter, file="3_output/offset_rasters/maxTemp_winter_offset.tif", format='GTiff', overwrite=TRUE)
```
 
#### Spring {-}

```{r eval=FALSE}
load("2_pipeline/store/models/brt_maxTemp_spring.rData")

p_spring <- predict(gee_all_spring, brt_maxTemp_spring, 
       n.trees=brt_maxTemp_spring$gbm.call$best.trees, type="response")

names(p_spring)<-"Tmax_diff_spring"

writeRaster(p_spring, file="3_output/offset_rasters/maxTemp_spring_offset.tif", format='GTiff', overwrite=TRUE)
```


### Difference between min temperature {.unnumbered}

#### Summer {-}

```{r eval=FALSE}
load("2_pipeline/store/models/brt_minTemp_summer.rData")

p_summer <- predict(gee_all_summer, brt_minTemp_summer,
       n.trees=brt_minTemp_summer$gbm.call$best.trees, type="response")

names(p_summer)<-"Tmin_diff_summer"

writeRaster(p_summer, file="3_output/offset_rasters/minTemp_summer_offset.tif", format='GTiff', overwrite=TRUE)
```

#### Fall {-}
```{r eval=FALSE}
load("2_pipeline/store/models/brt_minTemp_fall.rData")

p_fall <- predict(gee_all_fall, brt_minTemp_fall, 
       n.trees=brt_minTemp_fall$gbm.call$best.trees, type="response")

names(p_fall)<-"Tmin_diff_fall"

writeRaster(p_fall, file="3_output/offset_rasters/minTemp_fall_offset.tif", format='GTiff', overwrite=TRUE)
```


#### Winter {-}

```{r eval=FALSE}
load("2_pipeline/store/models/brt_minTemp_winter.rData")

p_winter <- predict(gee_all_winter, brt_minTemp_winter,
       n.trees=brt_minTemp_winter$gbm.call$best.trees, type="response")

names(p_winter)<-"Tmin_diff_winter"

writeRaster(p_winter, file="3_output/offset_rasters/minTemp_winter_offset.tif", format='GTiff', overwrite=TRUE)
```
 
#### Spring {-}

```{r eval=FALSE}
load("2_pipeline/store/models/brt_minTemp_spring.rData")

p_spring <- predict(gee_all_spring, brt_minTemp_spring, 
       n.trees=brt_minTemp_spring$gbm.call$best.trees, type="response")

names(p_spring)<-"Tmin_diff_spring"

writeRaster(p_spring, file="3_output/offset_rasters/minTemp_spring_offset.tif", format='GTiff', overwrite=TRUE)
```
